+++ src/testFixtures/kotlin/no/nav/syfo/common/mock/token/azuread/MockAzureAdTokenResponse.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.common.mock
+package no.nav.syfo.common.mock.token.azuread
 
 import com.fasterxml.jackson.annotation.JsonProperty